+++ kubejs/server_scripts/minecraft/recipes.js
@@ -1143,4 +1143,17 @@ const registerMinecraftRecipes = (event) => {
 			.duration(50)
 			.EUt(2)
 	})
+
+	// Greenhouse
+	generateGreenHouseRecipe(event, '8x minecraft:bamboo', '#tfc:any_fresh_water', 4000,
+		'64x minecraft:bamboo', 'bamboo', 'minecraft:overworld', 8, 
+		'8x minecraft:bamboo', GTValues.VA[GTValues.LV])
+
+	generateGreenHouseRecipe(event, '4x minecraft:red_mushroom', '#tfc:any_fresh_water', 4000,
+		'24x minecraft:red_mushroom', 'red_mushroom', 'minecraft:the_nether', 8, 
+		'4x minecraft:red_mushroom', GTValues.VA[GTValues.LV])
+
+	generateGreenHouseRecipe(event, '4x minecraft:brown_mushroom', '#tfc:any_fresh_water', 4000,
+		'24x minecraft:brown_mushroom', 'brown_mushroom', 'minecraft:the_nether', 8, 
+		'4x minecraft:brown_mushroom', GTValues.VA[GTValues.LV])
 }